create testing utility for multiple request
