+++ dev-packages/e2e-tests/test-applications/nuxt-4/nuxt.config.ts
@@ -5,11 +5,11 @@ export default defineNuxtConfig({
 
   routeRules: {
     '/rendering-modes/client-side-only-page': { ssr: false },
-    '/rendering-modes/pre-rendered-page': { prerender: true },
-    '/rendering-modes/swr-cached-page': { swr: true },
-    '/rendering-modes/swr-1h-cached-page': { swr: 3600 },
     '/rendering-modes/isr-cached-page': { isr: true },
     '/rendering-modes/isr-1h-cached-page': { isr: 3600 },
+    '/rendering-modes/swr-cached-page': { swr: true },
+    '/rendering-modes/swr-1h-cached-page': { swr: 3600 },
+    '/rendering-modes/pre-rendered-page': { prerender: true },
   },
 
   modules: ['@pinia/nuxt', '@sentry/nuxt/module'],
